requirements.txt
awscli-local==0.20
boto3 
psycopg2-binary
pandas
hashlib
